clients/mill-py/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mill-py"
version = "0.1.0"
description = "Python client for Mill data services — connect via gRPC or HTTP, query, and consume results as native Python, Arrow, pandas, or polars."
authors = ["qpointz <dev@qpointz.io>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/qpointz/qpointz"
repository = "https://github.com/qpointz/qpointz"
documentation = "https://github.com/qpointz/qpointz/tree/main/clients/mill-py"
keywords = ["mill", "grpc", "data", "client", "arrow", "pandas", "polars", "pyarrow"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database :: Front-Ends",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
packages = [{include = "mill"}]

# --------------------------------------------------------------------------
# Core dependencies (always installed)
# --------------------------------------------------------------------------
[tool.poetry.dependencies]
python = "^3.10,<4.0"   # 3.10 (negotiable), 3.11, 3.12 (firm), 3.13 (negotiable)
grpcio = "^1.60"
protobuf = "^5.27"
httpx = "^0.27"

# --------------------------------------------------------------------------
# Optional extras — DataFrame backends (all build on PyArrow)
# --------------------------------------------------------------------------
[tool.poetry.extras]
arrow = ["pyarrow"]
pandas = ["pyarrow", "pandas"]
polars = ["pyarrow", "polars"]
all = ["pyarrow", "pandas", "polars"]

# --------------------------------------------------------------------------
# Dev / test dependencies
# --------------------------------------------------------------------------
[tool.poetry.group.dev.dependencies]
grpcio-tools = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-mock = "*"
coverage = "*"
pytest-cov = "*"
pyarrow = "*"
pandas = "*"
polars = "*"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: unit tests (no server required, run by default)",
    "integration: integration tests (requires running Mill service)",
]
# By default run only unit tests; use `pytest -m integration` for integration
addopts = "-m unit"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["mill"]
omit = ["mill/_proto/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
